+++ project/src/store/api-actions.ts
@@ -9,9 +9,9 @@ import { OfferReviews, ReviewData } from '../types/offer-reviews';
 import { Offers, Offer } from '../types/offers';
 import { UserData } from '../types/user-data';
 import { redirectToRoute } from './action';
-import { loadNearbyOffers, loadOffer, loadReviews, sendReview } from './data-process/data-process';
-import { requireAuthorization } from './offer-process/offer-process';
-import { loadOffers } from './offers-process/offers-process';
+import { loadNearbyOffers, loadOffer, loadReviews, sendReview } from './offer-process/offer-process';
+import { requireAuthorization } from './user-process/user-process';
+import { loadOffers } from './main-process/main-process';
 
 
 /**
