+++ project/src/store/user-process/user-process.ts
@@ -1,0 +1,19 @@
+import {createSlice} from '@reduxjs/toolkit';
+import {NameSpace, AuthorizationStatus} from '../../const';
+import {UserProcess} from '../../types/state';
+
+const initialState: UserProcess = {
+  authorizationStatus: AuthorizationStatus.Unknown,
+};
+
+export const userProcess = createSlice({
+  name: NameSpace.user,
+  initialState,
+  reducers: {
+    requireAuthorization: (state, action) => {
+      state.authorizationStatus = action.payload;
+    },
+  },
+});
+
+export const {requireAuthorization} = userProcess.actions;
